refactored user profile
